CRUD purchase orders

+++ src/main/java/com/rapidtech/restapi/repository/PurchaseOrderRepo.java
@@ -3,5 +3,5 @@ package com.rapidtech.restapi.repository;
 import com.rapidtech.restapi.entity.PurchaseOrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseOrderRepo extends JpaRepository<PurchaseOrderEntity, Integer> {
+public interface PurchaseOrderRepo extends JpaRepository<PurchaseOrderEntity, Long> {
 }
